+++ src/messages/en/messages.js
@@ -31,6 +31,7 @@ module.exports = {
     dateBeforeToday: 'The :attribute must be a date before :arg0 :arg1.',
     dateFormat: 'The :attribute does not match the date format :arg0.',
     datetime: 'The :attribute must be a valid datetime(YYYY-MM-DD HH:mm:ss).',
+    dateiso: 'The :attribute must be a valid ISO-8601 date.',
     decimal: 'The :attribute must be a valid decimal value.',
     different: 'The :attribute and :arg0 must be different.',
     digits: 'The :attribute must be of :arg0 digits.',